Move gpgpdump.Options to options package,
