+++ options/options.go
@@ -1,4 +1,4 @@
-package gpgpdump
+package options
 
 import "fmt"
